Move AddressHooks to external package
