+++ external/bridges/parsers/addresshook.go
@@ -1,0 +1,7 @@
+package parsers
+
+import "strings"
+
+func DefaultLowercaseAddressHook(address string) (string, error) {
+	return strings.ToLower(address), nil
+}
